+++ CHANGELOG.md
@@ -14,6 +14,10 @@ project.
 
 ### Changed
 
+- The constrained DTW only tracks a single counter now.
+  By using the positive and the negative value range, we can save 1/3 of the overall RAM.
+  This should improve performance.
+
 ### Deprecated
 
 ### Removed